[FEATURE] Action Awareness: Show created file names
This feature tracks created files of session participants and
sends the file name to the other participants of the session. The
information is displayed in the new UIAwarenessTreeElement, introduced
by the new action awareness feature, see:

http://saros-build.imp.fu-berlin.de/gerrit/#/c/1572/

Additionally, the awareness information is printed in the multi user
chat and in the status line of the Eclipse workbench. This is done for
further user experiments to investigate, which place is the preferred
one for action awareness information. This can be configured in the
saros-properties file. Currently, just the roster is configured to
display the awareness information.

Change-Id: I01385e2126d681dedc80ac90fa767eab507b9a1d
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1829
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>