[INTERNAL] moved XMPP Feature parameter out of XMPPConnectionService
Due to changes in Smack 3.4.x it is now possible to add XMPP features to
a not connected connection. This patch makes the XMPPConnectionService
more generic as the features can now be added in the listener callback
and so the Saros XMPP feature support is added in the ConnectionHandler
instead.

Change-Id: I108e861863b240e9c68b90cee3f59ec0aa1925cc
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1860
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>