[INTERNAL] removed BlockProgressMonitor
The work
http://www.inf.fu-berlin.de/inst/ag-se/theses/Rintsch09-agile-tool-develop.pdf

states that some operations are performed asynchronous. I take a look at
the default implementations and they are NOT asynchronous. I also highly
doubt the other implementations will perform this asynchronous and
instead will pass the responsibility to the caller. From my own
experience my UI will freeze sometimes when my HDD is in standby mode
and has to spin again to fetch data so this is a clear indication for
synchronous processing.

Last but not least there is no evidence on his assumption, e.g at least
mention one provider that implements the behavior which is described in
this section.

Change-Id: Ia18437ab796580834a2b34344baea132a8444d0e
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1911
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>