[INTERNAL] make sure threads do not wait unnecessary
It currently happens that the OPN can perform idle waiting during user
list synchronization because the session is about to stop.

This patch does not fix anything but ensures that those threads will
terminate as fast as possible.

TODO: Find a way to ensure that this *signal* will be raised before the
session stops. The current problem is that all Negotiations will be
*asked* to stop before the session stops. As negotiations may be stuck
at this point due to user list synchronization they will still not
terminate before the session stops itself.

Change-Id: I78f79f0370824ee52494efce6470af8655035a13
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1914
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>