[INTERNAL] simplified FileUtils code
Removed dead backup method.

All resource operations will no longer try to force their operations and
so preventing data loss. Also removed the option to remove the read-only
flags when modifying data as there is mostly a reason such resources are
set to read only.

Also removed the whole workspace run stuff as it was mostly only one
file operation that was executed inside the runnable. The other multi
file operations are folder creation operations and those operations are
somewhat unlikely to be in conflict with other operations so there is
also no need to run them atomically.


Change-Id: Ia77ad3cff054c73101b7792d104ee199d9828547
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1919
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>