[INTERNAL] moved refresh logic from core to IDE logic
This behavior is heavy platform dependent. Although there is no need for
implementing such logic for specific IDE to the file-system interfaces
it does not offer any advantage to perform such logic in the core as it
may have unwanted side effects.

Put the logic to the Eclipse caller part of the affected methods. 

Change-Id: I8d82ebf2802f035962b078a52785ae1f5e433d3a
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1920
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>