[FEATURE] Adds a new item to the tab folder
This feature tracks the following user activities:
   * opening IDE dialogs
   * activating IDE views
   * starting and finishing test runs
   * executing refactorings
   * creating files

The information are displayed in a new tab item in addition to the
session and single user chats. Since this feature is still in
development, the information are just displayed as raw text in the tab
and in further development, this might be changed.

Change-Id: Id975b379434cda0e7b074ffbba378c5678909040
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1973
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>