[API|INTERNAL|REFACTOR] Refactored code and removed display from MUC, statusline and session overview
The action awareness feature was developed up to the change
http://saros-build.imp.fu-berlin.de/gerrit/#/c/1829

It contained the display of the action awareness information in three
different places:
   * MultiUserChat
   * Eclipse Status Line
   * Saros Session Overview
But user tests have shown that the action awareness information are
desired to be displayed at the place of the session chat. Instead of
mixing awareness information with chat lines, a new tab item will be
added to the tab folder in the Saros view. But for this purpose, the
code for the action awareness feature must be refactored.

This change removes the display at the three mentioned places, removes
some listeners to retrieve update information and also refactors some
pieces of code.

In another change after this, the display will be added. This change
just contains the logic to collect the awareness information and to
store them internally.

Change-Id: Ifd0a818b1f8d93c3ef74c24f447f387651f28eaa
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1975
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>