[FIX] Fixes some bugs in the retrieval of action awareness information
* Sometimes a view also has a description and was also displayed. Now,
just the tab's name is displayed.

* If the user switches between multiple editors, it was always
displayed. Now it'll be just displayed once.

* Some views were displayed twice. The reason was, that sometimes
dialogs are disposed which were not activated, e.g. 'Progress
Information'. After activating the Saros view during a session start,
this progress information dialog was disposed (without being seen) so
that the Saros as landing view was retrieved and in the end was
displayed twice in the activity log. The same occured when e.g. the
'Change Method Signature' or other refactoring dialogs were used, where
the display of the Editor as active view was displayed twice. Now, this
is fixed.

* Using runSafeSWTAsync in the handleEvent method in the ShellListener
caused problems in the order of the displayed activities. Now,
runSafeSWTSync is used, so that all activities are displayed in the
correct order.

* Using the Find/Replace dialog caused some problems (because it fired
the Deactivate event, which was removed). Now, it works.

Change-Id: I9f714683ab4a41181b2fd58c3929be8c7c484d14
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2003
Tested-by: Jenkins CI
Reviewed-by: Holger Schmeisky <google@schmeisky.com>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>