Revert '[BUILD] Merge branch development/actionAwareness into master'
This reverts commit 8b5f4979108de1f2bdc81d77e650507b9c015e97.
(Side note: This patch was created manually as there is no "revert" for
merge commits. I verified the accuracy of this patch by comparing it to
the last commit on the master branch before the merge, i.e. c8b0374. The
only difference between this patch and c8b0374 were, as expected, the
changes of commit 81a2041, which was applied after the merge.)

The action log feature seems to interfere heavily with other Saros
features, resulting in massive STF test case failures.

This might be a performance (and hence, timing issue), since the first
failing test was "AddMultipleFilesTest" (which creates 1000 small files
and hence 1000 additional action log entries).

This issue should be resolved on 'development/actionAwareness' and then
this branch can be reconsidered for merging.

Change-Id: I6be60121abe647a4f4f978e077513ca5e269cf04
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2018
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>