[API/BUILD] Adds SWT plugin for upcomming IntelliJ browser
It is planned to display the Saros UI in an embedded webbrowser in the future.
For this the SWT browser will be used, which was the result of the evaluation.
The evaluation result will be linked at http://www.inf.fu-berlin.de/w/SE/ThesisTestingSaros
in the near future.
As IntelliJ has no SWT libaries, they have to be distributed with Saros.
Due to licence conflicts (Saros under GPL vs. SWT under EPL) this has to be done
in a separate plugin.
This commit sets the basis for a web-based implementation by providing
the SWT libraries for all relevant OS as an IntelliJ plugin. The plugin has no
further responsibilities. The contained libraries are intended to be loaded
dynamically by the Saros/I plug-in. Therefore, the .iml is empty.

Change-Id: I6b833fbd9738a50c8f17c42a977b03fdd4edacd9
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2027
Tested-by: Jenkins CI
Reviewed-by: Holger Schmeisky <google@schmeisky.com>