[INTERNAL] Moves two XMPP related methods into ui package
These two methods were part of the ConnectServerAction class
which resides in the IntelliJ module. They are needed by the upcoming
HTML GUI to implement a minimalistic account management.
The classes of this commit are still work in progress and may be
moved into the core by either Matthias or Arndt as part of their theses.

Added new build.xml for Ui package and modified the build.xml of the IntelliJ
module.

Change-Id: I4cb42c7fece069d2eabd42c474b14e6e36b23c38
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2042
Tested-by: Jenkins CI
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>
Reviewed-by: Holger Schmeisky <google@schmeisky.com>