[UI] HTML-UI: adds rudimentary support to add accounts
This commit adds the HTML, CSS, Javascript, and Java files to display
a simple drop-down button to add and list XMPP accounts. A minimal HTML
dialog window supports the addition of new accounts.
All components are in a very early stage of development.
This commit adds the Angular JS library which will probably be removed
as soon the HTML leaves the prototype state.

Change-Id: Id4cd7c2e50370297a533cbaad8640ed6dd0f32ea
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2043
Tested-by: Jenkins CI
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>
Reviewed-by: Holger Schmeisky <google@schmeisky.com>