[INTERNAL] Moves SarosContext into core
This class was duplicated in both Eclipse and IntelliJ code.
The CTOR of SarosContext was changed to accept a list of
context factories instead of one factory with an optional
nested factory because the factories have no enclosing relationship,
they are on the same level. Furthermore, the code is now simpler.

Change-Id: I963464aa5931e921950dc002ab4185848876fe86
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2075
Tested-by: Jenkins CI
Reviewed-by: Christian Cikryt <c.cikryt@googlemail.com>