[INTERNAL] Replaced SWT browser with swt-improved library
The main advantage right now is that this library can delay the
execution of Javascript until a website is completely loaded.
The library is still work in process but added now to test its
compatibility inside both IntelliJ and Eclipse and to decide
which further features might be useful.

The project homepage of this library is here:
https://github.com/ag-se/swt-browser-improved

Change-Id: Ie72e6d375e833dee07e45b05f05ae2b342153192
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2108
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>