[INTERNAL] removed reinject method from SarosPluginContext class
This method offers the same functionality as the initComponent(...)
method.The side effect of adding an instance of a class to the current
context was never needed.

Removed this method as it is also error prone to use as it requires
exact knowledge when a specific component is added to the context.

Change-Id: I119fe580c17c5eee1f2f5f6f965886c23e6fd32c
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2080
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>