[INTERNAL|API] changed BrowserPage#getWebpage() IFSPEC
The current return value was to strict in such a way that it gives the
caller no clue about how to retrieve the physical content of a given
page. Furthermore this solution was far from flexible and may require
additional logic (e.g in an OSGi environment) for retrieving the
physical file location, i.e it is hard to tell what URLs are returned
from the implementations when running them in different environments.

To simplify this process only the absolute minimum of information is
returned that is needed by the caller to resolve such references by
itself.

Change-Id: I92a305d7dc0267e473a4261bf3c724dad6bf498f
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2093
Tested-by: Jenkins CI
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>