[INTERNAL] Let the different browser instances be managed by a class
There are different browser instances for the different dialogs and the main window
and those browsers change whenever a window is closed and re-opened.
The management of those instances is done by the new BrowserManager class.
As a result, the renderer classes don't have to be re-created each time the linked
browser changes and can now be managed by picocontainer.
Furthermore, they can now store the rendered state directly.

Change-Id: Ic24cd2fccfe2d7f36ad1f337a61391c8a6dbf853
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2119
Tested-by: Jenkins CI
Reviewed-by: Bastian Sieker <bsieker@mail.uni-paderborn.de>
Reviewed-by: Christian Cikryt <c.cikryt@googlemail.com>