[FIX][HTML] Timeout exception when browser view is invisible
If the HTML UI is activated, it will register a connection listener at startup.
Changing the connection state in the Saros SWT view while the
browser view is closed lead to a Timeout exception because the
render could not find the browser.
This commit changes the behaviour: the render call now checks if the browser
is present and does not wait for a timeout anymore. If the browser is
absent the render call is a NOP.

Change-Id: I0ad380acc2d54a80eccc00036322e5055268b857
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2147
Tested-by: Jenkins CI
Reviewed-by: Christian Cikryt <c.cikryt@googlemail.com>