[FIX] UI displayed incorrect Saros support
(Cherry-picked from master, was 9ff322e.)

During a refactor / design changes the original semantic on how to
select the imaged was changed. In the old design after a feature change
broadcast the DiscoveryManager was queried again when the Viewer updated
its content. This was overlooked when the RosterEntry class was changed
in such a way the those entries are now immutable.

Change-Id: Icd1b4d8e8132058416bbdf0ffa9a054f0d208835
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2194
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>