[FIX] UI displayed incorrect Saros support
During a refactor / design changes the original semantic on how to
select the imaged was changed. In the old design after a feature change
broadcast the DiscoveryManager was queried again when the Viewer updated
its content. This was overlooked when the RosterEntry class was changed
in such a way the those entries are now immutable.

Change-Id: I65d315a0aab5f4e1b5502a2c7be9440be4bbf0e5
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2174
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>