[INTERNAL] SharedResoucesManager triggered warning log message
When a client receives a FileActivity from type "Created",
the SharedResourcesManager tries to read the old content of this file
from the workspace to compare it with the FileActivity content.

If this file does not exist, the FileUtils class catch an Exception and
a warning occurs.

So every time a new file is created the logger shows an unnecessary
warning.

Original Author: David Damm

Change-Id: I3f81377d03ad706612ebec074292aa4087c33e06
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2175
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>