[INTERNAL] removed VCS usage for diff calculation on IPN side
The diff method of the FileList class has never used VCS meta
data when calculating the file diff but only the checksums.

Therefore this calculation is nothing more than a NOP that only adds
additional time to the (already "slowish") synchronization process when
VCS support is enabled.

Furthermore the whole diff is already calculated and the current process
contains defects anyways i.e:

Alice starts a session with Bob.
Afterwards Alices invites Carl.

While Carl is in the UI Wizard for selecting the target project either
Alice or Bob adds more files to the project that should be shared with
Carl.

Result:

When Carl has finished the negotiation the newly added files are
MISSING and can only be recovered if those files are opened in an
Eclipse editor window so that the watchdog notices the absence of these
files on Carls side.

Change-Id: I335421238d442ec4a2556a81cc3e825ed13fae0d
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2197
Tested-by: Jenkins CI
Reviewed-by: Arndt Lasarzik <alasarzik@gmail.com>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>