[HTML][UI] Adds StartSessionInvitation to UI
This commit adds a new UserInvitationPage Dialog with all needed
dependencies to display it the UI. This Dialog/Wizard should serve the
purpose of the "share with multiple contacts" wizard in the Eclipse SWT
Version of Saros. 

This commit includes a rudimentary HTML file but NO JS file updates,
since these are under heavily construction.  

Change-Id: I95f2ced8042f58b09c449a160098a031ad94ab8c
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2213
Tested-by: Jenkins CI
Reviewed-by: Arndt Lasarzik <alasarzik@gmail.com>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>