[HTML] Introduce new Javascript frontend essentials
A large part of this commit is the removal of the previous utilised
JavaScript frameworks which needs no further review.

Furthermore, the groundwork for the upcomming new GUI
based on HTML and JavaScript is provided.

IMPORTANT NOTE: For now, you have to build the JavaScript by
yourself to actually see changes in the application. Read the
README.md for further information.

Change-Id: I9c2d439195ca6da9d70f6730d6e57b574f31ebde
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2230
Tested-by: Jenkins CI
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>
Reviewed-by: Christian Cikryt <c.cikryt@googlemail.com>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>