[HTML] Add saros-state model, account dropdown and dictionary
A state model is introduced to manage UI state like the label of
the connect button etc.

An account model is introduced. Two corresponding views are
responsible to render them as a dropdown.

A dictionary is now responsible to provide strings for the UI to
avoid hard-coding them in different places (in JavaScript and or
JADE-templates).

Change-Id: I7a8cc996e76fefaad9494333210710cba007f109
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2233
Tested-by: Jenkins CI
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>