[REFACTOR] Rename CancelableProcess to Negotiation
The class is the superclass of all negotiation classes,
lives in de.fu_berlin.inf.ddp.negotiation, and would
be a great place for sharing methods that don't directly
relate to cancelability (such as collectPacket(), which
is currently duplicated between ProjectNegotiation and
SessionNegotiation). For these reasons, Negotiation seems
to be the most appropriate name.

Change-Id: Ia6341c969fad1d6882b1972075372564044a01a8
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2235
Tested-by: Jenkins CI
Reviewed-by: Bastian Sieker <bsieker@mail.uni-paderborn.de>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>