[HTML][UI] Introduces ProjectListRenderer
This Renderer uses an IDE independent IWorkspaceRoot to extract all
projects inside the workspace while creating the corresponding
ProjectTree models. Like most of the UI components this Renderer is
managed by PicoContainer to determinate what WorkspaceRoot
implementation is used.

Due to the missing support for circling references in JSON, the parent
attribute is removed from ProjectTreeNode model. A setType is added in
order to switch the node type after it's initialization. This commit
also includes some little improvements to the JDoc of the previously
introduced ProjectTree and ProjectTreeNode.

Change-Id: Ia5fed5d0f6bc3f8e2a818b2819f8c3a01b353958
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2237
Tested-by: Jenkins CI
Reviewed-by: Bastian Sieker <bsieker@mail.uni-paderborn.de>
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>