[HTML] Use bundle for multiple HTML webpages
There are multiple different webpages to show in Saros, for
example the main page, the incoming session negotiation
wizard, the outgoing ..., more wizards...

To determine which page to load (from Java, see package webpages)
the path to the html file must be provided. However, we want
to use the same JavaScript codebase for every page to exploit
code reusage. Furthermore, it is more convenient to have one
JavaScript build artefact for all pages instead of having
one artefact for every page.

With this patch, in the html file must be specified which page
view to show via setting the page property of the app object.

Change-Id: Ib5318f00c2d24aadad3727a75bd5ad41ed099a93
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2254
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>