[INTERNAL] Limits the amount of marshalled activities
This patch reduces the risk of getting an heap space error
by limiting the size of the list of activities.

The Xstream.marshal function uses a lot of heap memory while marshaling
the list. It raises by the size of the list.

Tests with a 30 MB large list results in a used
heap size of over 400 MB.

Change-Id: I7e8337fd42679011676fe36c27359c369984d312
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2251
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>