[INTERNAL] Adds WorkspaceRoot implementation to Eclipse
This commit change the IWorkspaceRoot interface to provide a list of
IProjects. This is used in the HTML UI to generate an IDE independent
ProjectModel. The EclipseImplementation of this is identical to the
Eclpise IWorkspaceRoot.getProjects() method. There is no IntelliJ
implementation yet.

This also add this implementation to SarosEclipseContextFactory so it
is manganed by picocontainer. This fixes the TODO:"Support
IWorkspaceRoot" in EclipseResourceImpl and ResourceAdapterFactory.

Change-Id: I181b151d84b1da7490c25192ad0e5ca1bd2f517b
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2236
Tested-by: Jenkins CI
Reviewed-by: Bastian Sieker <bsieker@mail.uni-paderborn.de>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>