[HTML] ProjectTree models on JavaScript side
Models for ProjectTree(s) and ProjectTreeNode(s) are introduced
to handle them on the JavaScript side.

Furthermore, a StartSessionWizard is added which will render
the contacts and projects for selection to start a session.

Further, this commit contains the some basic template files
to render selectable contacts and the project tree.

Change-Id: Id47d64b79b37379ca950f3abc05751c967c2e597
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2274
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>
Reviewed-by: Bastian Sieker <bsieker@mail.uni-paderborn.de>
Tested-by: Jenkins CI