[HTML][BUILD] Enhance JavaScript build and simplify build artefact
When building the JavaScript application via "npm run build", the
necessary directory structure is created and  all relevant files
are copied to it.

Furthermore, the build artefact of the ui.frontend package is
simplified: only the dist folder is extracted.

This change is related to
http://saros-build.imp.fu-berlin.de/gerrit/#/c/2258/ which
extracts bundle resources. The class EclipseResourceLocator
has to be adapted, accordingly.

Change-Id: Icd67effcb609d1c6b9fc2f48af72be3374bacc3b
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2272
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>