[HTML] Add jstree and simplify CSS embedding
To display the project list jstree (http://www.jstree.com/) is
added to the dependencies. The corresponding css classes and
image sprites are added, as well.

Instead of embedding the css files one by one in each html file,
now it is sufficient to embedd saros.css which loads all external
css files with the import statement.

Change-Id: I6e8c2371a362841784ba799ccfa76e19491546e0
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2275
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Denis Washington <denis@denisw.de>
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>