[HTML] BrowserPage#getTitle to show title in dialogs
"String getTitle()" function is added to the BrowserPage
interface. Its value will be shown as title in a dialog if the
BrowserPage is loaded in a dialog (the main-page for example is
not loaded in a dialog).

In addition, the size of a dialog is now set via setMinimumSize
instead of setSize since setSize was not working correctly.

Change-Id: Id43a049b22f42df4a599a721f859a0c11734cd42
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2293
Tested-by: Jenkins CI
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>
Reviewed-by: Holger Schmeisky <holger.schmeisky@gmail.com>