[INTERNAL][HTML] Introduces ProjectListManager
This adds a ProjectListManager to the UI. It is responsible for creating
and managing all ProjectTree models. It also provides functionality to
get the actual IProjects and IResources from a given ProjectTree model,
which will be used in the SessionInvitation.

The logic previously located in the ProjectListRenderer is now
transfered into this manager.

Change-Id: I3ca73a2d63519438572230d96d92dff24295f1b3
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2295
Tested-by: Jenkins CI
Reviewed-by: Bastian Sieker <bsieker@mail.uni-paderborn.de>
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>