[INTERNAL][HTML] Encapsulate application state in designated model
This commit reorganizes the handling and propagation of information
regarding the application state. The ContactListRenderer and the
ContactListFacade, which managed the list of contacts and the
connection state of the application (therefore, the name was
missleading) are replaced by the StateRenderer and the StateFacade.

Furthermore the ContactList model (the name was again missleading,
since it represented the list of contacts and the connection state)
is replaced with a State model which encapsulates the active
account, the corresponding contact list and the connection
state.

The ContactListManager is removed, instead, the StateRenderer
itself is responsible for listening to changes in the
ConnectionState and the Roster.

Change-Id: I1a23a4542f76d6c326917fa6c2cd613c362078e3
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2302
Tested-by: Jenkins CI
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>
Reviewed-by: Bastian Sieker <bsieker@mail.uni-paderborn.de>