[REFACTOR][E] Move getNickname out of User
This commit moves the getNickname from the User class to a new class
ModelFormatUtils that resides in the UI package. The nickname is only
relevant for UI components, so it should not be part of the network-layer
class User.
The new class UserFormatUtils manages retrieving the nickname from the
network lyer.
The new class ModelFormatUtils uses the session layer class UserFormatUtils
to get the nickname and provides a format() method to format model objects
 (e.g. users) for displaying in the UI (by nickname).

Along the way, some messages were extracted.

Change-Id: Idd15e05b637de9f437554afb48d36889d8cfe0c9
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2751
Reviewed-by: Holger Schmeisky <holger.schmeisky@gmail.com>
Tested-by: Jenkins CI