[INTERNAL][HTMl] Fixes for ProjectListManager
This fixes some problems with the JDoc and the naming of some variables
inside this class. In addition this class is now created in the
HTMLUIContextFactory. Furthermore instead of logging IOErrors in the
manager, they now are re-thrown and logged by the caller (in the moment
only ProjectListRenderer) to prevent multiple logs.

Change-Id: I926687e743a3061486dd046d0f38df1852dc6d13
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2763
Reviewed-by: Bastian Sieker <bsieker@mail.uni-paderborn.de>
Tested-by: Jenkins CI
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>