[API][HTML] Introduces ICollaborationUtils and IDE implementations
This commit introduces utils interface which offers convenient methods
for collaboration actions like sharing a project resources. Since this
Utils are already implemented in IntelliJ and Eclipse, the
implementation of this interfaces are just delegate the
method calls to these implementations. This class will be used in the
IDE independent HTML UI. 

Change-Id: I47f8d681bae003dcbacb051bdeca4f23668551f0
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2301
Tested-by: Jenkins CI
Reviewed-by: Bastian Sieker <bsieker@mail.uni-paderborn.de>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>