[INTERNAL][HTML] Adds session Invitation BrowserFunctions
This commit adds a JS browser function that allow it to start a session
with the HTML UI. The JS needs to be called with a GSON containing
ProjectTree models as first argument and a GSON containing a ContactList
model as second argument.

The actual resources are received thought the ProjcetListManager. The
session itself is started by a ICollaborationUtils

In order to use the collaboration utilities this adds 2 new methods to
the ProjectListManger. These are used to extract
only the resource for sharing without returning a mapping between a
project and his resources.

Change-Id: Ifaf16cda8df6f4e8de734556e00dbb7075202b6b
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2297
Reviewed-by: Bastian Sieker <bsieker@mail.uni-paderborn.de>
Tested-by: Jenkins CI
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>