[HTML] Contact contextmenu for rename and delete
An external small library is added to enable a contextmenu
on left-click. It is present in the file
`bootstrap-contextmenu.js` so you dont need to review this.

The contact view is extended by the contextmenu and buttons
for renaming and deleting a contact. For each of both
functionalities a dialog view is introduced.

The SarosApi is extended, accordingly.

Change-Id: I92a68f24eeb3078986b30590dbf784062f0ec815
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2773
Tested-by: Jenkins CI
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>