[Internal][HTML] Rework BrowserFunctions
This commit changes the whole BrowserFunctions (BFs from now on)creation
and seperation.

Previously the BFs were bundled for each BrowserPage usage. As a
single BF, like for example GetValidJID, can be used in many different
BrowserPages this bundling isn't reasonable anymore.

Instead of prebundling and wiring the BFs to single BrowserPages, each
BF is now created in a dedicated class inside the browser_functions
package. These classes have JDoc for better explanation of the purpose
and the JavaScript signature, which improves the usage of the previously
anonymous BF a lot. A HTML UI developer now has a place to look up the
name and semantic of injected JavaScript without searching big bundle
classes for the actual naming and implementation.

These BrowserFunctions and BrowserPages are injected 
field-injection. The BrowserPages and HTMLUIContext are changed
accordingly to this.

While reworking the BrowserFunctions this commit also adds a lot of
TODOs which should be done in later patches. 

Change-Id: I37bb9bcfe686b5890354af51d3511dce729dbec0
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2777
Tested-by: Jenkins CI
Reviewed-by: Bastian Sieker <bsieker@mail.uni-paderborn.de>
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>