[INTERNAL] Remove "newSession" flag from JoinSessionRequestExtension
This flag controlled whether the sender wants the receiver to
create a new session or invite the sender to an existing one.
This design is problematic because the sender doesn't know if
the receiver has already started a session; guessing wrong
(sending newSession=true if the receiver has started a session
already or newSession=false if there is no session yet) inevitably
causes an error response.

To solve this problem, this commit removes the newSession flag and
instead adopts the following logic:

* If there is no session running on the receiver when receiving the
  request, one is created.

* If there is a running session, the request's sender is invited to
  the existing session instead.

Change-Id: I29d1c11d253cd330c426d76b5c104a07c9c8a270
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2781
Tested-by: Jenkins CI
Reviewed-by: Denis Washington <denis@denisw.de>