[INTERNAL] Implement internalWorked() in CoreToEclipseMonitorAdapter
The former no-op implementation is replaced with one that accumulates
progress until full work units are reached and forwards those to
worked(). This makes the adapter work well with clients that call
internalWorked() directly instead of worked() (which seems to be the
case with some parts of Subclipse, for instance).

Change-Id: Ib42cddab9fbf77fcad2f213b1b47827ae0515f65
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2785
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>