[HTML] CSS styling
This commit makes some design changes driven by CSS.

Text-selection is disabled because that led to strange
visual effects, scrolling is now only allowed where it
should be (and Navigation is always "above" scrollable
content).

A content block (like the list of contacts and the
corresponding headline) are now visually sticked
together via a vertical bar.

Change-Id: Ifbe3e913458769370c237f693eb9243ccb877e47
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2790
Tested-by: Jenkins CI
Reviewed-by: Denis Washington <denis@denisw.de>
Reviewed-by: Bastian Sieker <bsieker@mail.uni-paderborn.de>