[REFACTOR][E] Free ConsistencyWatchdogServer from non-core dependencies
This makes use of the IEditorManager extensions added in previous
commits. Also required was a restructuring of DocumentChecksum, which
previously listened on its own for document changes with an (Eclipse)
IDocumentListener. Now, the ConsistencyWatchdogServer maintains a
single ISharedEditorListener for all documents to do this, which allowed
the DocumentChecksum class to be made much simpler.

Change-Id: I0866ad6e80855ef87ba12a1869ac200954cfe500
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2815
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>