[API] Introduce ISarosSessionContextFactory
This moves the creation of session components out of SarosSession and
into IDE-specific factory classes. It also allows components to be moved
to the core and be created in the SarosCoreSessionContextFactory that the
IDE-specific factories derive from. Both of these changes work towards
a later move of SarosSession to the core.

Change-Id: Iac28b7f38754da2c0c2572a7894344eb39318b4b
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2821
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>