[INTERNAL] Rework SarosSession.addSharedResources()
It was difficult to discern the four cases (1) new fully shared project,
(2) new partially shared project, (3) update partially shared to fully shared,
and (4) enlarge partially shared project.
This patch keeps the logic the same, but makes it easier to read.

Furthermore, the helper function "addMembers()" is reworked so the ugly field
"selectedResources" (aka the "WTF !!!" field) can be removed.

Change-Id: Icfd2b5a9a66e820cefc9cac6612ed5ce3df410b2
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2885
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>