[REFACTOR][E] Free ConsistencyWatchdogServer from non-core dependencies
This makes use of the IEditorManager extensions added in previous
commits. Also required was a restructuring of DocumentChecksum, which
previously listened on its own for document changes with an (Eclipse)
IDocumentListener. Now, the ConsistencyWatchdogServer maintains a single
ISharedEditorListener for all documents to do this, which allowed the
DocumentChecksum class to be made much simpler.

Change-Id: Icee9e79d6df76cb2c8244e6a903fde8fc902843b
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2840
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Tested-by: Jenkins CI