[FIX] #854 Jabber ID is correct but does not validate
JID validation now covers the most basic cases. IP literal
validation and length validation is still missing but both
cases should normally not occur in normal use cases as it is really
hard to exceed the length limit and IP literal usage for domain parts is
discouraged anyways.

It is still the question if Smack can handle this correctly or if the
JIDs have to be "punycoded" before they can be used with SMACK.

Change-Id: I98b8fed090fb3f08d696467574999bb52b071479
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2805
Tested-by: Jenkins CI
Reviewed-by: Denis Washington <denis@denisw.de>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>