[BUILD] Add libidn to Saros/I libraries
With cf36d4d, libidn was added to the core project.
But for the core project to compile in IntelliJ IDEA, it is not enough
to add an entry to de.fu_berlin.inf.dpp.core.iml: These entries only
refer to a name, not to a library itself:

<orderEntry type="library" exported="" name="libidn-1.32" level="project" />

What was missing is an entry in the libraries folder (in
de.fu_berlin.inf.dpp.intellij/.idea/libraries) that maps the library
name to a project-relativ path of a jar file.

Change-Id: Ic44bb8728899d74332d5d4b8d3112a8936935097
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2893
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Tested-by: Jenkins CI