[BUILD][I] Add Eclipse project information
Adds Eclipse .project, .classpath and .settings/ to the intellij project.
This makes it easier to import Saros/I in Eclipse for Refactorings of
the core.
Previously we considered to not include them, but since they also have to
be maintained (e.g. fest files and UI project  were missing), they should
 better be maintained here than in another place.

Change-Id: I629ecea91c0cc9481093cf9f6b77f2b9740bef45
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2896
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Tested-by: Jenkins CI
Reviewed-by: Holger Schmeisky <holger.schmeisky@gmail.com>