[REFACTOR] Avoid duplicate class names in same package
The two classes
 - de.fu_berlin.inf.dpp.test.util.MemoryPreferenceStore
 - de.fu_berlin.inf.dpp.test.util.TestThread
existed two times: one in the core project, the other in the Saros/E
project.

This is no problem on the CI server as the projects are independently
built, and the instances of the core are not visible to Saros/E.
In a developer's Eclipse workspace however, both classes are visible:
Saros/E JUnit tests, for instance, were provided with the core-version,
while they expected the Saros/E-version.

The core-version of MemoryPreferenceStore was introduced with 7610d3d.
The core-version of TestThread was introduced with 9a15f23.

This patch adds the prefix "Eclipse" in the Saros/E versions.
This refactoring could *not* be performed properly by the IDE: All
occurences where renamed first, mixing up core- and Saros/E-versions,
so I had to manually go through them.

Change-Id: I4135b3149d49496df60ce6b2af7868a5ec014039
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2892
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>