[API] Pass reason to ISessionLifecycleListener#sessionEnded
This enables us to remove the call to SarosView#showStopNotification
and replace it with an ISessionLifecycleListener in SarosView - which
in turn allows us to move LeaveAndKickHandler to the core.

The only non-automatic change made in this commit is the selection of
the SessionEndReason enum values to pass to sessionEnded() or the
stopSarosSession() method of SarosSessionManager.

Change-Id: I7e549f16e74d173c7faf5acb6f23dfb6f97260c5
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2902
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>
Reviewed-by: Denis Washington <denis@denisw.de>